Remove unnecessary test method separation
